+++ Support/Helpers/NumberFormatFunctions.ts
@@ -1,0 +1,1 @@
+export const dollarUSLocale = Intl.NumberFormat('en-US')