+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Next-Timehseet – Change Log
 
+### August 25th 
+
+- Refactored timesheet/page.ts to render min and max dates accurately
+- Refactored everhour/route.ts to build query-string params from object 
+- Refactored dates, array and number functions in support files
+- Added Support/Helpers folder
+
+
 ### August 24th
 
+- Added api/everhour routes
+- Created Components/WeekRoundUpBox component
+- Created timesheet page
 - First commit